rpn calc added

+++ stack.h
@@ -2,14 +2,14 @@
 #define STACK_H
 
 struct stack {
-    int val;
+    float val;
     int oper;
     struct stack *next;
 };
 
 typedef struct stack s;
 
-s *create_node(int val, int oper);
+s *create_node(float val, int oper);
 
 void append(s **head, s **tail, s **node);
